+++ src/main/resources/static/js/site/util.js
@@ -262,7 +262,6 @@ class Util {
         $(".takenAtDetails").text("");
         $(".manualTakenAtDetails").text("");
         $(".timeZoneDetails").text("");
-        $(".keywordsDetails").text("");
         $(".resolutionDetails").text("");
         $(".shareUrlDetails").html("");
         $(".coordinatesDetails").text("");
@@ -284,7 +283,6 @@ class Util {
         $(".takenAtLabel").hide();
         $(".manualTakenAtLabel").hide();
         $(".timeZoneLabel").hide();
-        $(".keywordsLabel").hide();
         $(".resolutionLabel").hide();
         $(".shareUrlLabel").hide();
         $(".coordinatesLabel").hide();
@@ -305,10 +303,6 @@ class Util {
         if (metadata.path != null) {
             $(".pathLabel").show();
             $(".pathDetails").text(metadata.path);
-        }
-        if (metadata.keywords != null) {
-            $(".keywordsLabel").show();
-            $(".keywordsDetails").text(metadata.keywords);
         }
         if (metadata.type != null) {
             $(".typeLabel").show();