+++ CHANGES.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Improved API response for unauthorized requests
+- Added keyword table and removed keyword field from metadata
 
 ### Fixed